refactor: restructure Trash module exports and update import paths
