+++ src/app/drive/components/DriveExplorer/components/DriveTopBarActions.tsx
@@ -11,7 +11,7 @@ import {
 } from '@phosphor-icons/react';
 import MoveActionIcon from 'assets/icons/move.svg?react';
 import { useSelector } from 'react-redux';
-import moveItemsToTrash from '../../../../../views/Trash/services/move-items-to-trash.service';
+import { moveItemsToTrash } from '../../../../../views/Trash/services';
 import errorService from '../../../../core/services/error.service';
 import navigationService from '../../../../core/services/navigation.service';
 import { DriveItemData, DriveItemDetails, FileViewMode } from '../../../../drive/types';
