+++ src/views/Trash/components/index.ts
@@ -1,0 +1,4 @@
+export { default as ClearTrashDialog } from './ClearTrashDialog';
+export { default as DeleteItemsDialog } from './DeleteItemsDialog';
+export { default as StopSharingAndMoveToTrashDialogWrapper } from './StopSharingAndMoveToTrashDialogWrapper';
+export { default as StopSharingAndMoveToTrashDialog } from './StopSharingAndMoveToTrashDialog';
