+++ src/views/Trash/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useTrashPagination } from './useTrashPagination';
