+++ src/views/Trash/services/batchProcessor.service.ts
