+++ src/views/Trash/services/clearTrash.service.ts
