+++ src/views/Trash/services/getTrash.service.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect, vi, beforeEach, Mock } from 'vitest';
-import { getTrash, getTrashPaginated, getWorkspaceTrashPaginated } from './get-trash.service';
+import { getTrash, getTrashPaginated, getWorkspaceTrashPaginated } from './getTrash.service';
 import { SdkFactory } from 'app/core/factory/sdk';
 import errorService from 'app/core/services/error.service';
 import localStorageService from 'app/core/services/local-storage.service';
