+++ src/views/Trash/services/index.ts
@@ -1,0 +1,4 @@
+export { getTrashPaginated, getWorkspaceTrashPaginated } from './getTrash.service';
+export { default as clearTrash } from './clearTrash.service';
+export { default as deleteItems } from './deleteItems.service';
+export { default as moveItemsToTrash } from './moveItemsToTrash.service';
